Refactor this out of enhanceCards to it's own module
